fix(impactanalysis) - moving classes back to the right package
