+++ src/main/java/com/synopsys/integration/detect/tool/impactanalysis/BlackDuckImpactAnalysisTool.java
@@ -33,6 +33,11 @@ import com.synopsys.integration.blackduck.service.model.ProjectVersionWrapper;
 import com.synopsys.integration.detect.configuration.DetectUserFriendlyException;
 import com.synopsys.integration.detect.configuration.enumeration.ExitCodeType;
 import com.synopsys.integration.detect.lifecycle.shutdown.ExitCodePublisher;
+import com.synopsys.integration.detect.tool.impactanalysis.service.ImpactAnalysis;
+import com.synopsys.integration.detect.tool.impactanalysis.service.ImpactAnalysisBatchOutput;
+import com.synopsys.integration.detect.tool.impactanalysis.service.ImpactAnalysisOutput;
+import com.synopsys.integration.detect.tool.impactanalysis.service.ImpactAnalysisUploadService;
+import com.synopsys.integration.detect.tool.impactanalysis.service.ImpactAnalysisUploadView;
 import com.synopsys.integration.detect.workflow.codelocation.CodeLocationNameManager;
 import com.synopsys.integration.detect.workflow.file.DirectoryManager;
 import com.synopsys.integration.detect.workflow.status.DetectIssue;
