+++ src/main/java/com/synopsys/integration/detect/tool/impactanalysis/ImpactAnalysisToolResult.java
@@ -15,6 +15,7 @@ import org.jetbrains.annotations.Nullable;
 
 import com.synopsys.integration.blackduck.codelocation.CodeLocationCreationData;
 import com.synopsys.integration.blackduck.service.model.NotificationTaskRange;
+import com.synopsys.integration.detect.tool.impactanalysis.service.ImpactAnalysisBatchOutput;
 
 public class ImpactAnalysisToolResult {
     @Nullable
